+++ AU/Public/Test-Package.ps1
@@ -1,5 +1,5 @@
 # Author: Miodrag Milic <miodrag.milic@gmail.com>
-# Last Change: 22-Oct-2016.
+# Last Change: 27-Oct-2016.
 
 <#
 .SYNOPSIS
@@ -33,9 +33,16 @@ function Test-Package {
         # Test chocolateyUninstall.ps1 only.
         [switch] $Uninstall,
 
+        # Path to chocolatey-test-environment to test package
+        [string] $Vagrant = $Env:au_Vagrant,
+
+        # Remove any other package in Vagrant 'packages' directory
+        [switch] $VagrantClear,
+
         # Package parameters
         $Parameters
     )
+
     if (!$Install -and !$Uninstall) { $Install = $Uninstall = $true }
 
     if (!$Nu) { $dir = gi $pwd }
@@ -66,6 +73,16 @@ function Test-Package {
     Write-Host "  Name:".PadRight(15)      $package_name
     Write-Host "  Version:".PadRight(15)   $package_version
     if ($Parameters) { Write-Host "  Parameters:".PadRight(15) $Parameters }
+    if ($Vagrant)    { Write-Host "  Vagrant: ".PadRight(15) $Vagrant }
+
+
+    if ($Vagrant) {
+        Write-Host "`nTesting package using vagrant"
+        if ($VagrantClear) { Write-Host 'Removing existing vagrant packages'; rm $Vagrant\packages\*.nupkg -ea ig }
+        cp $Nu $Vagrant\packages
+        start powershell -Verb Open -ArgumentList "-NoExit -Command `$Env:http_proxy=`$Env:https_proxy=`$Env:ftp_proxy=`$Env:no_proxy=''; cd $Vagrant; vagrant up --provision"
+        return
+    }
 
     if ($Install) {
         Write-Host "`nTesting package install"